requirements.txt
Django==2.1
django-debug-toolbar==1.9.1
django-widget-tweaks==1.4.2
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
sqlparse==0.2.4
